Phase 6: Fix LP64 struct layouts and host runtime for wasm64
Fix all struct field widths for LP64 ABI (sizeof(long)=8, sizeof(void*)=8):
- iovec: 16-byte entries (iov_base: usize, iov_len: usize)
- msghdr: 56-byte struct with pointer-width fields
- cmsghdr: 16-byte header (cmsg_len: size_t)
- stack_t: 24-byte struct (ss_sp: ptr, ss_size: size_t)
- sigaction: 24-byte struct (handler: usize)
- itimerval: 32 bytes (4 × i64 longs)
- mq_attr: long-sized fields

Add pointer-width helpers (read_usize_le/write_usize_le) for safe
struct field access. Update all function signatures using u32 for
pointer types to use usize (host_clone, host_futex, getsockname, etc.).

Host TypeScript: Fix BigInt conversions for memory64 (WebAssembly.Memory
with address:'i64', BigInt page counts, Number() wrappers for all kernel
export returns). Update CH_TOTAL_SIZE from 40+65536 to 72+65536.

Test runner scripts updated for wasm64-unknown-unknown target.

+++ crates/kernel/src/signal.rs
@@ -25,7 +25,7 @@ pub fn sig_bit(signum: u32) -> u64 {
 pub enum SignalHandler {
     Default,
     Ignore,
-    Handler(u32), // Function pointer (index) in guest Wasm -- for future use
+    Handler(usize), // Function pointer (index) in guest Wasm
 }
 
 /// Full sigaction information: handler + flags + mask.
